fixed stove point sync

+++ kitchen_feud/Assets/Scripts/Minigame/StoveMinigame/stoveMinigame.cs
@@ -57,7 +57,7 @@ public class stoveMinigame : MonoBehaviour
                 else if (stoveCanvas.tag == "Team2" && (int)PhotonNetwork.LocalPlayer.CustomProperties["Team"] == 1)
                     dishOfFoundDish.points = dishOfFoundDish.points * 2;
                 
-                dishOfFoundDish.GetComponent<PhotonView>().RPC("pointSync", RpcTarget.Others, dishOfFoundDish.points);
+                dishOfFoundDish.GetComponent<PhotonView>().RPC("pointSync", RpcTarget.Others, (int)dishOfFoundDish.points);
 
                 Debug.Log("UpdateDishPoints: " + dishOfFoundDish.points);
             }
